Extract 'HTML to Spanned' functionality to separate class
