+++ app/core/src/main/java/com/fsck/k9/message/html/HtmlConverter.kt
@@ -1,14 +1,8 @@
 package com.fsck.k9.message.html
 
 
-import android.text.Editable
-import android.text.Html
-import android.text.Html.TagHandler
-import android.text.Spanned
-
 import com.fsck.k9.K9
 import org.jsoup.Jsoup
-import org.xml.sax.XMLReader
 
 /**
  * Contains common routines to convert html to text and vice versa.
@@ -102,48 +96,5 @@ object HtmlConverter {
     @JvmStatic
     fun textToHtmlFragment(text: String): String {
         return TextToHtml.toHtmlFragment(text)
-    }
-
-    /**
-     * Convert HTML to a [Spanned] that can be used in a [android.widget.TextView].
-     *
-     * @param html The HTML fragment to be converted.
-     *
-     * @return A [Spanned] containing the text in `html` formatted using spans.
-     */
-    @JvmStatic
-    fun htmlToSpanned(html: String): Spanned {
-        return Html.fromHtml(html, null, ListTagHandler())
-    }
-
-    /**
-     * [TagHandler] that supports unordered lists.
-     *
-     * @see HtmlConverter.htmlToSpanned
-     */
-    private class ListTagHandler : TagHandler {
-        override fun handleTag(opening: Boolean, tag: String, output: Editable, xmlReader: XMLReader) {
-            if (tag == "ul") {
-                if (opening) {
-                    var lastChar: Char = 0.toChar()
-                    if (output.length > 0) {
-                        lastChar = output[output.length - 1]
-                    }
-                    if (lastChar != '\n') {
-                        output.append("\r\n")
-                    }
-                } else {
-                    output.append("\r\n")
-                }
-            }
-
-            if (tag == "li") {
-                if (opening) {
-                    output.append("\t•  ")
-                } else {
-                    output.append("\r\n")
-                }
-            }
-        }
     }
 }
