+++ app/ui/src/main/java/com/fsck/k9/activity/setup/WelcomeMessage.java
@@ -8,15 +8,19 @@ import android.view.View;
 import android.view.View.OnClickListener;
 import android.widget.TextView;
 
+import com.fsck.k9.DI;
 import com.fsck.k9.ui.R;
 import com.fsck.k9.activity.Accounts;
 import com.fsck.k9.activity.K9Activity;
-import com.fsck.k9.message.html.HtmlConverter;
+import com.fsck.k9.ui.helper.HtmlToSpanned;
+
 
 /**
  * Displays a welcome message when no accounts have been created yet.
  */
 public class WelcomeMessage extends K9Activity implements OnClickListener{
+    private static final HtmlToSpanned htmlToSpanned = DI.get(HtmlToSpanned.class);
+
 
     public static void showWelcomeMessage(Context context) {
         Intent intent = new Intent(context, WelcomeMessage.class);
@@ -30,7 +34,7 @@ public class WelcomeMessage extends K9Activity implements OnClickListener{
         setLayout(R.layout.welcome_message);
 
         TextView welcome = findViewById(R.id.welcome_message);
-        welcome.setText(HtmlConverter.htmlToSpanned(getString(R.string.accounts_welcome)));
+        welcome.setText(htmlToSpanned.convert(getString(R.string.accounts_welcome)));
         welcome.setMovementMethod(LinkMovementMethod.getInstance());
 
         findViewById(R.id.next).setOnClickListener(this);
